Attach and detach a Node to/from a Network
Adding a node to a network happens in two steps: First the node is added into the network's internal list of nodes and then the node's attach method is called.
Removing a node from a network happens in two steps: First the node's detach method is called and then the node is removed from the network's internal list of nodes.