Callbacks for Service
Implemented a callback system to the Service class.
Each callback should be as short as possible.

Use the methods add_event and remove_event to add or remove event names to the service.
Use the methods add_callback and remove_callback to add or remove callback to an event.
Use the method notify to call the callbacks for an event. Currently notify calls the callbacks directly.